requirements.txt
appdirs==1.4.3
django~=1.8.0
django-extensions==1.5.7
djangorestframework==3.2.3
dj-database-url==0.4.1
gunicorn==19.3.0
hashids==1.1.0
packaging==16.8
psycopg2==2.6.1
pyparsing==2.2.0
python-dotenv==0.1.3
six==1.10.0
two1==3.10.6
